modify: 캐스팅 오류 수정
- JwtAuthenticationFilter: jwt에서 user-id 값 읽어올 때 int로 먼저 읽어오도록 수정

+++ src/main/java/gist/pilldispenser/common/security/jwt/JwtAuthenticationFilter.java
@@ -50,7 +50,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
     }
 
     private UserDetails getMemberFromToken(String token) {
-        Long id = (long)jwtProvider.getClaimValue(token, "user-id");
+        Long id = (long)((int)jwtProvider.getClaimValue(token, "user-id"));
         return usersDetailsService.loadUserById(id);
     }
